backend/requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
